Move module tests to subfolders
